% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_create_core.R
\name{zar_create_core}
\alias{zar_create_core}
\title{Create a new core directory scaffold}
\usage{
zar_create_core(core, path, if_exists = "error")
}
\arguments{
\item{core}{Core directory name, e.g. "GKUT25_02".}

\item{path}{Path to the site directory where the core should be created.}

\item{if_exists}{What to do if core directory already exists: "error" or "skip".}
}
\value{
Invisibly returns the path to the created core directory.
}
\description{
Create a new core directory scaffold
}
