% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_hsitools.R
\name{zar_hsitools}
\alias{zar_hsitools}
\title{Install HSItools from GitHub}
\usage{
zar_hsitools(branch = "dev")
}
\arguments{
\item{branch}{Character. Which branch to install. Default \code{"dev"}.}
}
\value{
\code{NULL}, invisibly. Called for its side effect.
}
\description{
Install HSItools from GitHub
}
