% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_install.R
\name{zar_install}
\alias{zar_install}
\title{Install Maury's choice of packages with pak}
\usage{
zar_install()
}
\value{
\code{NULL}, invisibly.
}
\description{
Install Maury's choice of packages with pak
}
