% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_template_swir_coregister.R
\name{zar_template_swir_coregister}
\alias{zar_template_swir_coregister}
\title{Use SWIR coregister template}
\usage{
zar_template_swir_coregister(
  path,
  capture = NULL,
  vnir_capture,
  if_exists = "error"
)
}
\arguments{
\item{path}{Destination path (file or directory)}

\item{capture}{Capture directory name. If NULL and path is directory,
inferred from path.}

\item{vnir_capture}{Paired VNIR capture directory name. Used to locate
target GCPs and RGB reference raster from VNIR spatial directory.}

\item{if_exists}{What to do if file exists: "error", "skip", or "overwrite"}
}
\description{
Use SWIR coregister template
}
