% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_templates.R
\name{zar_template_swir_postprocess}
\alias{zar_template_swir_postprocess}
\title{Use SWIR postprocess template}
\usage{
zar_template_swir_postprocess(
  path,
  capture = NULL,
  vnir_capture,
  if_exists = "error"
)
}
\arguments{
\item{path}{Destination path (file or directory)}

\item{capture}{Capture directory name. If NULL and path is directory, inferred from path.}

\item{vnir_capture}{Paired VNIR capture directory name. Used to locate the
transect layer computed during VNIR postprocessing.}

\item{if_exists}{What to do if file exists: "error", "skip", or "overwrite"}
}
\description{
Use SWIR postprocess template
}
