% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_templates.R
\name{zar_template_vnir_postprocess}
\alias{zar_template_vnir_postprocess}
\title{Use VNIR postprocess template}
\usage{
zar_template_vnir_postprocess(path, capture = NULL, if_exists = "error")
}
\arguments{
\item{path}{Destination path (file or directory)}

\item{capture}{Capture directory name. If NULL and path is directory, inferred from path.}

\item{if_exists}{What to do if file exists: "error", "skip", or "overwrite"}
}
\description{
Use VNIR postprocess template
}
