% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_hsitools.R
\name{zar_hsitools}
\alias{zar_hsitools}
\title{Install HSItools from GitHub}
\usage{
zar_hsitools(branch = "dev")
}
\arguments{
\item{branch}{which branch to use. Default "dev".}
}
\description{
Install HSItools from GitHub
}
