% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_templates.R
\name{zar_template_swir}
\alias{zar_template_swir}
\title{Use SWIR reflectance template}
\usage{
zar_template_swir(path, overwrite = FALSE)
}
\arguments{
\item{path}{Destination path for the script}

\item{overwrite}{Overwrite existing file? Default FALSE}
}
\description{
Use SWIR reflectance template
}
