% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_templates.R
\name{zar_template_swir}
\alias{zar_template_swir}
\title{Use SWIR reflectance template}
\usage{
zar_template_swir(
  path,
  capture = NULL,
  reference = capture,
  if_exists = "error"
)
}
\arguments{
\item{path}{Destination path (file or directory)}

\item{capture}{Capture directory name. If NULL and path is directory, inferred from path.}

\item{reference}{Reference directory name. Defaults to capture.}

\item{if_exists}{What to do if file exists: "error", "skip", or "overwrite"}
}
\description{
Use SWIR reflectance template
}
