% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zar_templates.R
\name{zar_template_vnir}
\alias{zar_template_vnir}
\title{Use VNIR reflectance template}
\usage{
zar_template_vnir(path, overwrite = FALSE)
}
\arguments{
\item{path}{Destination path for the script}

\item{overwrite}{Overwrite existing file? Default FALSE}
}
\description{
Use VNIR reflectance template
}
